+++ src/main/java/helltech/dynamodb/model/Entity.java
