+++ src/main/java/helltech/dynamodb/model/Publication.java
