+++ src/main/java/helltech/dynamodb/persistence/dynamodb/DynamoDBRepository.java
@@ -1,17 +1,17 @@
-package helltech.dynamodb;
+package helltech.dynamodb.persistence;
 
-import static helltech.dynamodb.DatabaseConstants.GSI1;
-import static helltech.dynamodb.DatabaseConstants.GSI2;
-import static helltech.dynamodb.DatabaseConstants.GSI3;
-import static helltech.dynamodb.DatabaseConstants.TABLE_NAME;
-import helltech.dynamodb.model.business.Entity;
-import helltech.dynamodb.model.business.Institution;
-import helltech.dynamodb.model.business.Publication;
-import helltech.dynamodb.model.business.User;
-import helltech.dynamodb.model.dao.Dao;
-import helltech.dynamodb.model.dao.InstitutionDao;
-import helltech.dynamodb.model.dao.PublicationDao;
-import helltech.dynamodb.model.dao.UserDao;
+import static helltech.dynamodb.persistence.DynamoDbConstants.GSI1;
+import static helltech.dynamodb.persistence.DynamoDbConstants.GSI2;
+import static helltech.dynamodb.persistence.DynamoDbConstants.GSI3;
+import static helltech.dynamodb.persistence.DynamoDbConstants.TABLE_NAME;
+import helltech.dynamodb.model.Entity;
+import helltech.dynamodb.model.Institution;
+import helltech.dynamodb.model.Publication;
+import helltech.dynamodb.model.User;
+import helltech.dynamodb.persistence.dao.Dao;
+import helltech.dynamodb.persistence.dao.InstitutionDao;
+import helltech.dynamodb.persistence.dao.PublicationDao;
+import helltech.dynamodb.persistence.dao.UserDao;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
