+++ src/main/java/helltech/dynamodb/persistence/dynamodb/dao/PublicationDao.java
@@ -1,16 +1,16 @@
-package helltech.dynamodb.model.dao;
+package helltech.dynamodb.persistence.dao;
 
-import static helltech.dynamodb.DatabaseConstants.GSI2;
-import static helltech.dynamodb.DatabaseConstants.GSI3;
-import static helltech.dynamodb.DatabaseConstants.PK2;
-import static helltech.dynamodb.DatabaseConstants.PK3;
-import static helltech.dynamodb.DatabaseConstants.SK2;
-import static helltech.dynamodb.DatabaseConstants.SK3;
+import static helltech.dynamodb.persistence.dynamodb.DynamoDbConstants.GSI2;
+import static helltech.dynamodb.persistence.dynamodb.DynamoDbConstants.GSI3;
+import static helltech.dynamodb.persistence.dynamodb.DynamoDbConstants.PK2;
+import static helltech.dynamodb.persistence.dynamodb.DynamoDbConstants.PK3;
+import static helltech.dynamodb.persistence.dynamodb.DynamoDbConstants.SK2;
+import static helltech.dynamodb.persistence.dynamodb.DynamoDbConstants.SK3;
 import static software.amazon.awssdk.enhanced.dynamodb.TableSchema.fromBean;
 import helltech.dynamodb.annotations.Generated;
-import helltech.dynamodb.model.business.Institution;
-import helltech.dynamodb.model.business.Publication;
-import helltech.dynamodb.model.business.User;
+import helltech.dynamodb.model.Institution;
+import helltech.dynamodb.model.Publication;
+import helltech.dynamodb.model.User;
 import java.util.Objects;
 import java.util.UUID;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
